Add `Exports.Window` APIs for window state, properties, and navigation
- Implement `Get` and `Set` methods in both native and managed layers for window state and properties, including size, position, fullscreen, and transparency.
- Add native navigation APIs: `NavigateToString`, `NavigateToUrl`, and `SendWebMessage`.
- Introduce additional actions like `Center`, `Restore`, `ClearBrowserAutoFill`, and `ShowNotification`.
- Update `CMakeLists.txt` with new source files.
