+++ src/InfiniFrame.NativeBridge/Native/CMakeLists.txt
@@ -45,7 +45,10 @@ set(COMMON_SOURCES
         src/Api/Exports/Exports.Monitors.cpp
         src/Api/Exports/Exports.Platform.MacOs.cpp
         src/Api/Exports/Exports.Platform.Windows.cpp
-        src/Api/Exports/Exports.Window.cpp
+        src/Api/Exports/Exports.Window.Actions.cpp
+        src/Api/Exports/Exports.Window.Getters.cpp
+        src/Api/Exports/Exports.Window.Navigation.cpp
+        src/Api/Exports/Exports.Window.Setters.cpp
 )
 
 set(TEST_SOURCES
